requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-import-export==2.5.0
django-jazzmin==2.4.7
django-ranged-response==0.2.0
djangoql==0.14.5
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.1.5
Pillow==8.2.0
ply==3.11
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
